cleanup documentation

+++ CipedTronicMcu/CipedTronicMCU/Platform/EEprom.c
@@ -11,22 +11,22 @@
 #include "EEProm.h"
 
 
-void EEPROMWrite(uint8_t address, uint8_t data)
+void EEPROMWrite(uint16_t address, uint8_t data)
 {
 	eeprom_write_byte((uint8_t *)address, data);
 }
 
-uint8_t EEPROMRead(uint8_t address)
+uint8_t EEPROMRead(uint16_t address)
 {
 	return eeprom_read_byte ((const uint8_t*)address);	
 }
 
-void EEPROMWrite32(uint32_t address, uint32_t data)
+void EEPROMWrite32(uint16_t address, uint32_t data)
 {
 	eeprom_write_dword((uint32_t *)address, data);
 }
 
-uint32_t EEPROMRead32(uint32_t address)
+uint32_t EEPROMRead32(uint16_t address)
 {
 	return eeprom_read_dword ((const uint32_t*)address);	
 }
@@ -36,7 +36,7 @@ void EEPROMWriteChar(uint16_t address, char* data,int32_t len)
 	int16_t count = 0;
 	for(count = address;count < len + address;count++)
 	{
-		EEPROMWrite(address,(uint8_t)data[count]);
+		EEPROMWrite(address,data[count]);
 	}
 }
